Agrega nuevos cambios de reestructuración de carpetas
