+++ ibichos-dashboard/src/pages/Catalogo/catalogo.tsx
@@ -1,172 +1,20 @@
-import { useEffect, useState } from 'react';
-import { collection, onSnapshot, doc, updateDoc, setDoc } from 'firebase/firestore';
-import { db } from '../config/firebaseConfig';
-import { Search, Bug, MapPin, ShieldCheck, Trash2, Edit3, ShieldAlert } from 'lucide-react';
-import { useAuth } from '../context/authcontext';
-
-import type { Captura } from '../types/captura';
+// src/pages/Catalogo/Catalogo.tsx
+import { Search, Bug, MapPin, ShieldCheck, Trash2, ShieldAlert } from 'lucide-react';
+import { useCatalogo } from './useCatalogo';
 
 export default function Catalogo() {
-  const { user } = useAuth();
-  const [capturas, setCapturas] = useState<Captura[]>([]);
-  const [cargando, setCargando] = useState(true);
-  const [busqueda, setBusqueda] = useState('');
-  const [selectedImg, setSelectedImg] = useState<string | null>(null);
-  const [showModal, setShowModal] = useState(false);
-  const [targetId, setTargetId] = useState<string | null>(null);
-  const [userMap, setUserMap] = useState<Record<string, { name: string; email: string }>>({});
-
-  useEffect(() => {
-    import('firebase/firestore').then(({ getDocs, collection: col }) => {
-      getDocs(col(db, 'users')).then(snap => {
-        const map: Record<string, { name: string; email: string }> = {};
-        snap.forEach(d => {
-          map[d.id] = { name: d.data().displayName || 'Sin nombre', email: d.data().email || '' };
-        });
-        setUserMap(map);
-      });
-    });
-  }, []);
-
-  useEffect(() => {
-    setCargando(true);
-    // Escucha en tiempo real de la colección "captures"
-    const unsubscribe = onSnapshot(collection(db, "captures"), (querySnapshot) => {
-      const datos: Captura[] = querySnapshot.docs
-        .map(doc => {
-          const d = doc.data();
-          // Lógica inclusiva: Es aprobada si dice APPROVED 
-          // O si la IA tiene confianza alta (>0.75) y no está marcada para revisión
-          const currentStatus = d.status || d.validationStatus;
-          const isApproved = currentStatus === 'APPROVED' || 
-                            (!currentStatus && !d.needsReview && (d.probability || 0) >= 0.75);
-          
-          return { id: doc.id, ...d, isApproved, status: currentStatus } as any;
-        })
-        .filter(cap => cap.isApproved && cap.status !== 'DELETED')
-        .map(cap => ({
-          id: cap.id,
-          imageUrl: cap.imageUrl || '',
-          category: cap.category || 'OTHER',
-          dangerLevel: cap.dangerLevel || 'UNKNOWN',
-          confidence: cap.probability || 0,
-          userId: cap.userId || 'Anónimo',
-          insectName: cap.insectName,
-          scientificName: cap.scientificName,
-          status: cap.status || 'APPROVED',
-          timestamp: cap.timestamp // El timestamp ya se estaba extrayendo bien aquí
-        }));
-
-      // ORDENAMIENTO CRONOLÓGICO (Más reciente a más antiguo)
-      datos.sort((a, b) => {
-        const timeA = a.timestamp?.toDate ? a.timestamp.toDate().getTime() : new Date(a.timestamp || 0).getTime();
-        const timeB = b.timestamp?.toDate ? b.timestamp.toDate().getTime() : new Date(b.timestamp || 0).getTime();
-        
-        return timeB - timeA;
-      });
-
-      setCapturas(datos);
-      setCargando(false);
-    }, (error) => {
-      console.error("Error en tiempo real (Catálogo):", error);
-      setCargando(false);
-    });
-
-    return () => unsubscribe();
-  }, []);
-
-  const traducirPeligro = (level: string) => {
-    switch(level) {
-      case 'HARMLESS': return 'Inofensivo';
-      case 'VENOMOUS': return 'Venenoso';
-      case 'CAUTION': return 'Precaución';
-      default: return 'Desconocido';
-    }
-  };
-
-  const handleUpdate = async (id: string, updates: any) => {
-    try {
-      const docRef = doc(db, 'captures', id);
-      await updateDoc(docRef, {
-        ...updates,
-        moderatedBy: user?.uid || 'ADMIN',
-        moderatorEmail: user?.email || 'Admin',
-        moderatedAt: new Date()
-      });
-
-      try {
-        const actionType = updates.category ? 'UPDATE_CATEGORY' : (updates.dangerLevel ? 'UPDATE_DANGER_LEVEL' : 'UPDATE_CAPTURE');
-        const logRef = doc(collection(db, 'moderation_logs'));
-        await setDoc(logRef, {
-          adminId: user?.uid || 'unknown',
-          adminEmail: user?.email || 'Admin',
-          action: actionType,
-          targetId: id,
-          targetType: 'CAPTURE',
-          timestamp: new Date()
-        });
-      } catch (e) {
-        console.warn("No se pudo guardar el log en Catálogo:", e);
-      }
-
-      setCapturas(prev => prev.map(cap => cap.id === id ? { ...cap, ...updates } : cap));
-    } catch (error) {
-      console.error("Error al actualizar especie:", error);
-    }
-  };
-
-  const handleReject = async () => {
-    if (!targetId) return;
-    try {
-      const docRef = doc(db, 'captures', targetId);
-      await updateDoc(docRef, { 
-        status: 'REJECTED',
-        needsReview: false,
-        moderatedBy: user?.uid || 'ADMIN',
-        moderatorEmail: user?.email || 'Admin',
-        moderatedAt: new Date()
-      });
-
-      try {
-        const logRef = doc(collection(db, 'moderation_logs'));
-        await setDoc(logRef, {
-          adminId: user?.uid || 'unknown',
-          adminEmail: user?.email || 'Admin',
-          action: 'REJECT_CAPTURE', // O 'HIDE_CAPTURE' según convenga
-          targetId: targetId,
-          targetType: 'CAPTURE',
-          timestamp: new Date()
-        });
-      } catch (e) {
-        console.warn("No se pudo guardar el log en Catálogo:", e);
-      }
-
-      setCapturas(prev => prev.filter(cap => cap.id !== targetId));
-      setShowModal(false);
-      setTargetId(null);
-    } catch (error) {
-      console.error("Error al rechazar especie:", error);
-    }
-  };
-
-  const openDeleteModal = (id: string) => {
-    setTargetId(id);
-    setShowModal(true);
-  };
-
-  const getDangerBadge = (level: string) => {
-    switch(level) {
-      case 'HARMLESS': return 'bg-success';
-      case 'VENOMOUS': return 'bg-danger';
-      case 'CAUTION': return 'bg-warning text-dark';
-      default: return 'bg-secondary';
-    }
-  };
-
-  const filtered = capturas.filter(cap => 
-    cap.category.toLowerCase().includes(busqueda.toLowerCase()) ||
-    cap.id.toLowerCase().includes(busqueda.toLowerCase())
-  );
+  const {
+    cargando,
+    busqueda, setBusqueda,
+    selectedImg, setSelectedImg,
+    showModal, setShowModal,
+    userMap,
+    filtered,
+    handleUpdate,
+    handleReject,
+    openDeleteModal,
+    getDangerBadge
+  } = useCatalogo();
 
   if (cargando) {
     return (
@@ -315,6 +163,7 @@ export default function Catalogo() {
           </div>
         </>
       )}
+      
       {/* MODAL DE CONFIRMACIÓN (DISEÑO PREMIUM) */}
       {showModal && (
         <>
