+++ ibichos-dashboard/src/types/auditlog.ts
@@ -1,0 +1,9 @@
+export interface AuditLog {
+  id: string;
+  adminId: string;
+  adminEmail?: string;
+  action: string;
+  targetId: string;
+  targetType: string;
+  timestamp: Date;
+}
